REFACTOR: Renamed atom folder to ui as it is the standards
