+++ app/(site)/components/ProjectSection.tsx
@@ -1,6 +1,6 @@
 import validateSlugWithProject from "@/actions/projects/validateSlugWithProject";
-import Button from "@/components/Atoms/Button";
-import SlideUp from "@/components/Atoms/Slideup";
+import Button from "@/components/UI/Button";
+import SlideUp from "@/components/UI/Slideup";
 import ProjectItem from "@/components/ProjectItem/ProjectItem";
 import HeadingTwo from "@/components/Text/HeadingTwo";
 import allProjects from "@/constants/projects";
