+++ app/blogs/components/BlogListSection.tsx
@@ -1,4 +1,4 @@
-import Grid from "@/components/Atoms/Grid";
+import Grid from "@/components/UI/Grid";
 import BlogItem from "@/components/Blogs/BlogItem";
 import HeadingTwo from "@/components/Text/HeadingTwo";
 import { Blog } from "@/types/blog";
