+++ app/blogs/page.tsx
@@ -1,7 +1,7 @@
 import HeadingOne from "@/components/Text/HeadingOne";
 import { BlogList } from "./components/BlogList";
 import type { Metadata } from "next";
-import PageDescription from "@/components/Atoms/PageDescription";
+import PageDescription from "@/components/UI/PageDescription";
 import blogs from "@/constants/blogs";
 
 const description = `
