+++ app/skills/[slug]/components/CredentialsSection.tsx
@@ -1,5 +1,5 @@
 import CredentialListSection from "@/app/credentials/components/CredentialListSection";
-import Button from "@/components/Atoms/Button";
+import Button from "@/components/UI/Button";
 import Certificate from "@/types/certificates";
 import { Skill } from "@/types/skills";
 import Link from "next/link";
