+++ app/skills/[slug]/components/ProjectsSection.tsx
@@ -1,5 +1,5 @@
 import ProjectsListSection from "@/app/projects/components/ProjectListSection";
-import Button from "@/components/Atoms/Button";
+import Button from "@/components/UI/Button";
 import Project from "@/types/projects";
 import { Skill } from "@/types/skills";
 import Link from "next/link";
